Time: 3 ms (29.61%) | Memory: 8.5 MB (54.38%) - LeetSync

+++ 7-reverse-integer/reverse-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int reverse(int x) {
-        long long int r = 0;
+        long int r = 0;
         while(x!=0){
             r = r*10 + x%10;
             x = x/10;
